chore: organiza arquivos
